requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
arrow==1.3.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.0.4
backoff==2.2.1
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
croniter==1.4.1
cryptography==38.0.4
databricks-sql-connector==2.0.3
dataclasses==0.6
duckdb==0.3.4
duckdb-engine==0.1.11
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
frozenlist==1.4.1
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gql==3.5.0
graphql-core==3.2.3
graphviz==0.18.2
greenlet==3.0.3
h11==0.14.0
halo==0.0.31
idna==3.6
imageio==2.34.0
imutils==0.5.4
iniconfig==2.0.0
Jinja2==3.1.3
jsonschema==3.2.0
lazy_loader==0.3
log-symbols==0.0.14
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
metricflow-lite==0.130.2
mo-dots==4.22.21108
mo-future==3.147.20327
mo-imports==3.149.20327
mo-kwargs==4.22.21108
mo-logs==4.23.21108
monotonic==1.6
more-itertools==8.10.0
moz-sql-parser==4.40.21126
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
opencv-contrib-python ==4.9.0.80
opencv-python==4.9.0.80
oscrypto==1.3.0
packaging==23.2
pandas==1.4.4
pathspec==0.12.1
pillow==10.2.0
pillow-heif==0.16.0
pluggy==1.4.0
protobuf==4.25.3
pyarrow==5.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pycron==3.0.0
pycryptodomex==3.20.0
pydantic==1.10.14
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==22.1.0
pyrsistent==0.20.0
pytesseract == 0.3.10
pytest==8.0.2
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
rudder-sdk-python==1.0.6
scikit-image==0.22.0
scipy==1.12.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowflake-connector-python==2.9.0
snowflake-sqlalchemy==1.2.3
spinners==0.0.24
SQLAlchemy==1.4.51
sqlalchemy-bigquery==1.9.0
sqlalchemy-redshift==0.8.1
sqlalchemy2-stubs==0.0.2a38
starlette==0.36.3
tabulate==0.8.9
termcolor==2.4.0
tesserocr-Pillow == 4.1.1 
thrift==0.13.0
tifffile==2024.2.12
tomli==2.0.1
transform==1.4.3
transform-tools==1.1.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
update-checker==0.18.0
urllib3==1.26.18
uvicorn==0.27.1
validators==0.22.0
xxhash==3.4.1
yamllint==1.35.1
yarl==1.9.4
